+++ package.json
@@ -21,6 +21,7 @@
   "homepage": "https://github.com/sims226/tweet-parser#readme",
   "dependencies": {
     "dotenv": "^4.0.0",
+    "express": "^4.15.3",
     "twitter": "^1.7.1"
   }
 }
