+++ public/index.html
@@ -14,7 +14,9 @@
       <h2>Enter a Twitter handle:</h2>
       <div id="input-container">
         <input type="text" id="user-input" maxlength="25">
-        <button id="grabInput" onclick="storeInput()">Submit</button>
+        <!--<button id="grabInput" onclick="storeInput()">Submit</button>-->
+        <button id="grabInput">Submit</button>
+        
       </div>
     </div>
     <div id="results-container">
@@ -59,6 +61,6 @@
 
       <p>Fast-forward a couple of decades and I find myself learning to code. This page, then, is both an indulgence to my own practice and a little homage to one of my favorite authors.</p>
     </div>
-    <script src="app.js"></script>
+    <!--<script src="app.js"></script>-->
   </body>
 </html>
